+++ lib/federails.rb
@@ -6,6 +6,7 @@ require 'federails/version'
 require 'federails/engine'
 require 'federails/configuration'
 require 'federails/utils/object'
+require 'federails/utils/json_request'
 
 # rubocop:disable Style/ClassVars